consumer-service/requirements.txt
dnspython==2.4.2
pika==1.3.2
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
